# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Claw Machine Backend API
  version: 1.0.0
servers:
  - url: /
paths:
  /v1/auth/telegram:
    post:
      summary: Telegram auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData]
              properties:
                initData:
                  type: string
      responses:
        '201':
          description: Auth success
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresInSec, user]
                properties:
                  accessToken:
                    type: string
                  expiresInSec:
                    type: integer
                  user:
                    type: object
                    required: [id, telegramUserId]
                    properties:
                      id:
                        type: string
                      telegramUserId:
                        type: string
  /v1/auth/dev:
    post:
      summary: Development auth (non-Telegram)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                devUserId:
                  type: string
      responses:
        '201':
          description: Dev auth success
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresInSec, user]
                properties:
                  accessToken:
                    type: string
                  expiresInSec:
                    type: integer
                  user:
                    type: object
                    required: [id, telegramUserId]
                    properties:
                      id:
                        type: string
                      telegramUserId:
                        type: string
  /v1/attempts/start:
    post:
      summary: Start attempt
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: { type: string }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machineId, clientBuild, configVersion]
              properties:
                machineId: { type: string }
                clientBuild: { type: string }
                configVersion: { type: string }
      responses:
        '201':
          description: Attempt started
          content:
            application/json:
              schema:
                type: object
                required: [attemptId, attemptToken, serverNowMs, inputWindowMs, economySnapshot]
                properties:
                  attemptId: { type: string }
                  attemptToken: { type: string }
                  serverNowMs: { type: integer }
                  inputWindowMs: { type: integer }
                  economySnapshot:
                    type: object
                    required: [ticketsLeft]
                    properties:
                      ticketsLeft: { type: integer }
  /v1/attempts/{attemptId}/inputs:
    post:
      summary: Send attempt inputs
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
        - in: header
          name: Authorization
          required: true
          schema: { type: string }
        - in: header
          name: X-Attempt-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [packets]
              properties:
                packets:
                  type: array
                  items:
                    type: object
                    required: [seq, clientTimeMs, moveX, moveY]
                    properties:
                      seq: { type: integer }
                      clientTimeMs: { type: integer }
                      moveX: { type: number }
                      moveY: { type: number }
      responses:
        '201':
          description: Inputs accepted
          content:
            application/json:
              schema:
                type: object
                required: [acceptedSeqUpTo, serverNowMs, warnings]
                properties:
                  acceptedSeqUpTo: { type: integer }
                  serverNowMs: { type: integer }
                  warnings:
                    type: array
                    items: { type: string }
  /v1/attempts/{attemptId}/resolve:
    post:
      summary: Resolve attempt
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
        - in: header
          name: Authorization
          required: true
          schema: { type: string }
        - in: header
          name: X-Attempt-Token
          required: true
          schema: { type: string }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientSummary]
              properties:
                clientSummary:
                  type: object
                  required: [pressTimeMs]
                  properties:
                    pressTimeMs: { type: integer }
                    closeStartMs: { type: integer }
                    contactHints:
                      type: array
                      items:
                        type: object
                        required: [toyHintId, fingers]
                        properties:
                          toyHintId: { type: string }
                          fingers: { type: integer }
      responses:
        '201':
          description: Attempt resolved
          content:
            application/json:
              schema:
                type: object
                required: [attemptId, status, result, riskScore]
                properties:
                  attemptId: { type: string }
                  status:
                    type: string
                    enum: [resolved]
                  result:
                    type: string
                    enum: [win, lose, void]
                  reward:
                    type: object
                    properties:
                      id: { type: string }
                      code: { type: string }
                      rarity: { type: string }
                  seedReveal: { type: string }
                  riskScore: { type: integer }
  /v1/rewards/claim:
    post:
      summary: Claim reward
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: { type: string }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attemptId]
              properties:
                attemptId: { type: string }
      responses:
        '201':
          description: Claim state
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [granted, already_granted, pending, failed]
                  reward:
                    type: object
                    properties:
                      code: { type: string }
                      rarity: { type: string }